requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
confuse==2.0.1
dnspython==2.7.0
idna==3.10
load-dotenv==0.1.0
pymongo==3.12.3
PyMySQL==1.1.1
python-amazon-sp-api==1.6.86
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
urllib3==2.2.3
